+++ test/end-to-end/helpers/NightmareHelper.js
@@ -1,7 +1,3 @@
-/**
- * Created by doug on 24/04/18.
- */
-
 'use strict';
 
 const Helper = codecept_helper;
@@ -10,9 +6,8 @@ class NightmareHelper extends Helper {
 
     clickBackBrowserButton() {
         const browser = this.helpers['Nightmare'].browser;
-        console.log('pre-browser-back');
+
         return browser.back();
-        console.log('post-browser-back');
     }
 
 }
