--- test/end-to-end/helpers/RESTHelper.js
@@ -1,12 +1,0 @@
-let request = require('unirest');
-
-class RESTHelper extends codecept_helper {
-    doHttpGet (url) {
-        return request.get(url)
-            .then((body) => {
-                return body;
-            });
-    }
-}
-
-module.exports = RESTHelper;
