+++ test/end-to-end/pages/coapplicant/startPage.js
@@ -5,6 +5,6 @@ module.exports = function () {
     const I = this;
 
     I.seeCurrentUrlEquals(pageUnderTest.getUrl());
-    I.click('.button.button-start');
+    I.click('.button');
 
 };
